docker-app/requirements.txt
streamlit>=1.22.0
boto3>=1.28.0
pandas
matplotlib
plotly
markdown
python-dotenv
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
mcp>=0.1.0
